+++ NEWS.md
@@ -1,5 +1,6 @@
-### httptest 1.0.1 (under development)
+## httptest 1.1.0
 
+* `expect_header` to assert that a HTTP request has a header
 * Always prune the leading ":///" that appears in `with_mock_API` if the URL has a querystring
 
 # httptest 1.0.0